==> 03_feature_engineering.ipynb <==
import gc
import time
import numpy as np
import pandas as pd
from datetime import datetime

### Example for extend the features

d1 = {'card_id':[1, 2, 1, 3], 
      'A':[1, 2, 1, 2], 
      'B':[2, 1, 2, 2], 
      'C':[4, 5, 1, 5], 
      'D':[7, 5, 4, 8],}

numeric_cols = ['C', 'D']
category_cols = ['A', 'B']

t1 = pd.DataFrame(d1)
t1

features = {}
card_all = t1['card_id'].values.tolist()
for card in card_all:
    features[card] = {}
features

columns = t1.columns.tolist()
columns

idx = columns.index('card_id')
category_cols_index = [columns.index(col) for col in category_cols]
numeric_cols_index = [columns.index(col) for col in numeric_cols]
print(category_cols_index, numeric_cols_index)


# Combine categorical and continuous features
for i in range(t1.shape[0]):
    va = t1.loc[i].values
    card = va[idx]
    for cate_ind in category_cols_index:
        for num_ind in numeric_cols_index:
            print(columns[cate_ind], va[cate_ind], columns[num_ind])
            col_name = '&'.join([columns[cate_ind], str(va[cate_ind]), columns[num_ind]])
            features[card][col_name] = features[card].get(col_name, 0) + va[num_ind]
features

df = pd.DataFrame(features).T.reset_index()
cols = df.columns.tolist()
df.columns = ['card_id'] + cols[1:]
df

This method of feature creation can efficiently capture hidden information in the dataset. However, it tends to generate a significant number of missing values, which should be considered in the subsequent modeling process due to the potential issue of a sparse feature matrix.

# Generate features combination

train = pd.read_csv('./dataset/preprocess/train_pre.csv')
test =  pd.read_csv('./dataset/preprocess/test_pre.csv')
transaction = pd.read_csv('./dataset/preprocess/transaction_d_pre.csv')

numeric_cols = ['purchase_amount', 'installments']

category_cols = ['authorized_flag', 'city_id', 'category_1',
       'category_3', 'merchant_category_id','month_lag','most_recent_sales_range',
                 'most_recent_purchases_range', 'category_4',
                 'purchase_month', 'purchase_hour_section', 'purchase_day']

id_cols = ['card_id', 'merchant_id']

# 创建字典用于保存数据
features = {}
card_all = train['card_id'].append(test['card_id']).values.tolist()
for card in card_all:
    features[card] = {}
     
# 标记不同类型字段的索引
columns = transaction.columns.tolist()
idx = columns.index('card_id')
category_cols_index = [columns.index(col) for col in category_cols]
numeric_cols_index = [columns.index(col) for col in numeric_cols]

# 记录运行时间
s = time.time()
num = 0

# 执行循环，并在此过程中记录时间
for i in range(transaction.shape[0]):
    va = transaction.loc[i].values
    card = va[idx]
    for cate_ind in category_cols_index:
        for num_ind in numeric_cols_index:
            col_name = '&'.join([columns[cate_ind], str(va[cate_ind]), columns[num_ind]])
            features[card][col_name] = features[card].get(col_name, 0) + va[num_ind]
    num += 1
    if num%1000000==0:
        print(time.time()-s, "s")
del transaction
gc.collect()

# 字典转dataframe
df = pd.DataFrame(features).T.reset_index()
del features
cols = df.columns.tolist()
df.columns = ['card_id'] + cols[1:]

# 生成训练集与测试集
train = pd.merge(train, df, how='left', on='card_id')
test =  pd.merge(test, df, how='left', on='card_id')
del df
train.to_csv("./dataset/preprocess/train_dict.csv", index=False)
test.to_csv("./dataset/preprocess/test_dict.csv", index=False)

gc.collect()

# Generate features based on statistical results

We can generate new featuers by their statistical results, such as sum, mean, varience. In this way, the new features are not likely contain so many NAN values.

transaction = pd.read_csv('./dataset/preprocess/transaction_g_pre.csv')

numeric_cols = ['authorized_flag',  'category_1', 'installments',
       'category_3',  'month_lag','purchase_month','purchase_day','purchase_day_diff', 'purchase_month_diff',
       'purchase_amount', 'category_2', 
       'purchase_month', 'purchase_hour_section', 'purchase_day',
       'most_recent_sales_range', 'most_recent_purchases_range', 'category_4']
categorical_cols = ['city_id', 'merchant_category_id', 'merchant_id', 'state_id', 'subsector_id']

aggs = {}

for col in numeric_cols:
    aggs[col] = ['nunique', 'mean', 'min', 'max', 'var', 'skew', 'sum']
for col in categorical_cols:
    aggs[col] = ['nunique']
aggs['card_id'] = ['size', 'count']
cols = ['card_id']

for key in aggs.keys():
    cols.extend([key+'_'+stat for stat in aggs[key]])

df = transaction[transaction['month_lag'] < 0].groupby('card_id').agg(aggs).reset_index()
df.columns = cols[:1] + [co+'_hist' for co in cols[1:]]

df2 = transaction[transaction['month_lag'] >= 0].groupby('card_id').agg(aggs).reset_index()
df2.columns = cols[:1] + [co+'_new' for co in cols[1:]]
df = pd.merge(df, df2, how='left', on='card_id')

df2 = transaction.groupby('card_id').agg(aggs).reset_index()
df2.columns = cols
df = pd.merge(df, df2, how='left',on='card_id')
del transaction
gc.collect()

train = pd.merge(train, df, how='left', on='card_id')
test =  pd.merge(test, df, how='left', on='card_id')
del df
train.to_csv("./dataset/preprocess/train_groupby.csv", index=False)
test.to_csv("./dataset/preprocess/test_groupby.csv", index=False)

gc.collect()

# Merge two kinds of new features

Up to this point, we have completed the feature extraction from two different perspectives. However, as of now, the features from the two sets of solutions are still stored in separate data files. We need to merge them to proceed with modeling. The merging process is relatively straightforward; you just need to horizontally concatenate "train_dict" (or "test_dict") and "train_group" (or "test_group") based on the "card_id" column and then remove any duplicate columns.

train_dict = pd.read_csv("./dataset/preprocess/train_dict.csv")
test_dict = pd.read_csv("./dataset/preprocess/test_dict.csv")
train_groupby = pd.read_csv("./dataset/preprocess/train_groupby.csv")
test_groupby = pd.read_csv("./dataset/preprocess/test_groupby.csv")

for co in train_dict.columns:
    if co in train_groupby.columns and co!='card_id':
        del train_groupby[co]
for co in test_dict.columns:
    if co in test_groupby.columns and co!='card_id':
        del test_groupby[co]

train_dict.head()

Missing values will be filled with 0 in this case. It's important to note that these missing values are not truly missing but rather represent values that were not accounted for during the feature creation process. These values can be logically interpreted as 0. Therefore, filling missing values with 0 here is essentially a data completion step.

train = pd.merge(train_dict, train_groupby, how='left', on='card_id').fillna(0)
test = pd.merge(test_dict, test_groupby, how='left', on='card_id').fillna(0)

train.to_csv("./dataset/preprocess/train.csv", index=False)
test.to_csv("./dataset/preprocess/test.csv", index=False)

del train_dict, test_dict, train_groupby, test_groupby
gc.collect()
