==> 04_train_and_fine_tuning.ipynb <==
import gc
import time
import numpy as np
import pandas as pd
from datetime import datetime
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV

# Random Forest

train = pd.read_csv("./dataset/preprocess/train.csv")
test = pd.read_csv("./dataset/preprocess/test.csv")

## Feature selection

features = train.columns.tolist()
features.remove('card_id')
features.remove('target')
featureSelect = features[:]

corr = []
for fea in featureSelect:
    corr.append(abs(train[[fea, 'target']].fillna(0).corr().values[0][1]))

se = pd.Series(corr, index=featureSelect).sort_values(ascending=False)
feature_select = ['card_id'] + se[:300].index.tolist()

train = train[feature_select + ['target']]
test = test[feature_select]

## Grid Search

features = train.columns.tolist()
features.remove('card_id')
features.remove('target')

parameter_space = {
    "n_estimators": [79, 80, 81], 
    "min_samples_leaf": [29, 30, 31],
    "min_samples_split": [2, 3],
    "max_depth": [9, 10],
    "max_features": ["auto", 80]
}

clf = RandomForestRegressor(
    criterion="squared_error",
    n_jobs=15,
    random_state=22)

grid = GridSearchCV(clf, parameter_space, cv=2, scoring="neg_mean_squared_error")
grid.fit(train[features].values, train['target'].values)

grid.best_estimator_

np.sqrt(-grid.best_score_)

grid.best_estimator_.predict(test[features])

test['target'] = grid.best_estimator_.predict(test[features])
test[['card_id', 'target']].to_csv("./result/submission_randomforest.csv", index=False)
