==> 05_cross_validation.ipynb <==
import os
import numpy as np
import pandas as pd
import gc
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import GridSearchCV, KFold
from numpy.random import RandomState

train = pd.read_csv('./dataset/preprocess/train.csv')
test = pd.read_csv('./dataset/preprocess/test.csv')

train.shape

train.head()

np.count_nonzero(train) / train.size

features = train.columns.tolist()
features.remove('card_id')
features.remove('target')
featureSelect = features[:]

corr = []
for fea in featureSelect:
    corr.append(abs(train[[fea, 'target']].fillna(0).corr().values[0][1]))

se = pd.Series(corr, index=featureSelect).sort_values(ascending=False)
feature_select = ['card_id'] + se[:300].index.tolist()

train_RF = train[feature_select + ['target']]
test_RF = test[feature_select]

train_RF.shape

test_RF.shape

def feature_select_pearson(train, test):
    """
    Feature selection based on the pearson coefficient
    :param train:  training set
    :param test: test set
    :return: training and test sets after selection
    """
    print('feature_select...')
    features = train.columns.tolist()
    features.remove("card_id")
    features.remove("target")
    featureSelect = features[:]

    # 去掉缺失值比例超过0.99的
    for fea in features:
        if train[fea].isnull().sum() / train.shape[0] >= 0.99:
            featureSelect.remove(fea)

    # 进行pearson相关性计算
    corr = []
    for fea in featureSelect:
        corr.append(abs(train[[fea, 'target']].fillna(0).corr().values[0][1]))

    # 取top300的特征进行建模，具体数量可选
    se = pd.Series(corr, index=featureSelect).sort_values(ascending=False)
    feature_select = ['card_id'] + se[:300].index.tolist()
    print('done')
    return train[feature_select + ['target']], test[feature_select]

def param_grid_search(train):
    """
    网格搜索参数调优
    :param train:训练集
    :return:网格搜索训练结果
    """

    print('param_grid_search')
    features = train.columns.tolist()
    features.remove("card_id")
    features.remove("target")
    parameter_space = {
        "n_estimators": [81], 
        "min_samples_leaf": [31],
        "min_samples_split": [2],
        "max_depth": [10],
        "max_features": [80]
    }
    

    print("Tuning hyper-parameters for mse")

    clf = RandomForestRegressor(
        criterion="squared_error",
        n_jobs=15,
        random_state=22)

    grid = GridSearchCV(clf, parameter_space, cv=2, scoring="neg_mean_squared_error")
    grid.fit(train[features].values, train['target'].values)
    

    print("best_params_:")
    print(grid.best_params_)
    means = grid.cv_results_["mean_test_score"]
    stds = grid.cv_results_["std_test_score"]
    # 此处额外考虑观察交叉验证过程中不同超参数的
    for mean, std, params in zip(means, stds, grid.cv_results_["params"]):
        print("%0.3f (+/-%0.03f) for %r"
              % (mean, std * 2, params))
    return grid

grid = param_grid_search(train_RF)

np.sqrt(-grid.best_score_)

# Cross validation

def train_predict(train, test, best_clf):
    """
    To train and predict
    :param train: training set
    :param test: test set
    :param best_clf: the best classifier
    :return:
    """

    # step 1: feature selection
    print('train_predict...')
    features = train.columns.tolist()
    features.remove("card_id")
    features.remove("target")

    # step 2: Cross Validation
    prediction_test = 0
    cv_score = []
    prediction_train = pd.Series()

    kf = KFold(n_splits=5, random_state=22, shuffle=True)
    for train_part_index, eval_index in kf.split(train[features], train['target']):
        # train the model on training set
        best_clf.fit(train[features].loc[train_part_index].values, train['target'].loc[train_part_index].values)
        # Add the prediction result
        prediction_test += best_clf.predict(test[features].values)
        # prediction on validation test
        eval_pre = best_clf.predict(train[features].loc[eval_index].values)
        # evaluate on validation prediction with MSE
        score = np.sqrt(mean_squared_error(train['target'].loc[eval_index].values, eval_pre))
        # Put MSE score into cv_score list
        cv_score.append(score)
        print(score)
        # Put prediction on validation set into prediction_train
        prediction_train = prediction_train.append(pd.Series(best_clf.predict(train[features].loc[eval_index]), index=eval_index))
    
    print(cv_score, sum(cv_score)/5)
    pd.Series(prediction_train.sort_index().values).to_csv("./dataset/preprocess/train_randomforest.csv", index=False)
    pd.Series(prediction_test / 5).to_csv("./dataset/preprocess/test_randomforest.csv", index=False)
    # Append the label 'target' to the test set
    test['target'] = prediction_test / 5
    test[['card_id', 'target']].to_csv("result/submission_randomforest_5kf.csv", index=False)
    return

train_predict(train_RF, test_RF, grid.best_estimator_)
